OpenAPI virtual mock service
