+++ modules/swagger-codegen/src/main/resources/JavaSpring/api.mustache
@@ -15,6 +15,7 @@ import io.swagger.mock.MockUtil;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import io.swagger.mock.model.MockRequest;
 import io.swagger.mock.model.MockResponse;
+import io.swagger.mock.model.MockServiceRequest;
 import java.util.HashMap;
 import java.util.Map;
 {{/generateDynamicMock}}
