+++ modules/swagger-codegen/src/main/resources/JavaSpring/apiController.mustache
@@ -28,6 +28,7 @@ import org.springframework.web.multipart.MultipartFile;
 {{#generateDynamicMock}}
 //Customized for MOCK
 import io.swagger.mock.MockUtil;
+import io.swagger.mock.model.MockServiceRequest;
 import io.swagger.mock.model.MockRequest;
 import io.swagger.mock.model.MockResponse;
 import org.apache.commons.lang3.builder.EqualsBuilder;
