+++ modules/swagger-codegen/src/main/resources/JavaSpring/apiControllerMockLogic.mustache
@@ -1,26 +1,27 @@
 {{#hasParams}}	
-Map<String, String> paramMap =  new HashMap<>();
-	{{#allParams}}
-		{{#isQueryParam}}
-		if({{paramName}} != null && {{paramName}}.getClass().toString().contains("List")) {
-			paramMap.put("{{baseName}}", addQueryParamValue({{paramName}}));					
-		} else {
-			paramMap.put("{{baseName}}", String.valueOf({{paramName}}));
-		}
-		{{/isQueryParam}}
-		{{#isPathParam}}
+		Map<String, String> paramMap =  new HashMap<>();
+		MockServiceRequest mockServiceRequest = new MockServiceRequest();
+{{#allParams}}
+	{{#isQueryParam}}
+	if({{paramName}} != null && {{paramName}}.getClass().toString().contains("List")) {
+		paramMap.put("{{baseName}}", addQueryParamValue({{paramName}}));					
+	} else {
+		paramMap.put("{{baseName}}", String.valueOf({{paramName}}));
+	}
+	{{/isQueryParam}}
+	{{#isPathParam}}
 		paramMap.put("{{baseName}}", String.valueOf({{paramName}}));{{/isPathParam}}
 	{{/allParams}}
 		try {
-			mockUtil.setResource("{{baseName}}");
-			mockUtil.setOperationId("{{operationId}}");
-			mockUtil.setParams(paramMap);
+			mockServiceRequest.setResource("{{baseName}}");
+			mockServiceRequest.setOperationId("{{operationId}}");
+			mockServiceRequest.setParams(paramMap);
 			{{#bodyParams}}
-			mockUtil.setInputObjectType({{#allParams}}{{>bodyParamsInput}}{{/allParams}}.getClass());
-			mockUtil.setInputObject({{#allParams}}{{>bodyParamsInput}}{{/allParams}});
+			mockServiceRequest.setInputObjectType({{#allParams}}{{>bodyParamsInput}}{{/allParams}}.getClass());
+			mockServiceRequest.setInputObject({{#allParams}}{{>bodyParamsInput}}{{/allParams}});
 			{{/bodyParams}}
-			return mockUtil.returnResponse();
-		} catch (Exception e) {
+			return mockUtil.returnResponse(mockServiceRequest);
+		} catch (ClassNotFoundException | IOException e) {
 			log.error("Unable to load the mock Response for " + "{{operationId}}", e);
 			return new ResponseEntity("{\"code\": \"ERROR\", \"message\": \"Unable to load the mock Response for {{operationId}}\"}", HttpStatus.INTERNAL_SERVER_ERROR);
 		}
